Readded fixed date check, removed whole section of tests
DD.MM-DD or DD-DD.MM are formats that are really easy to mess up and they are hard to read. Even tests that were passing to this day were wrong, as the month in the only day part would always default to current month, so the end part could be several months behind. There just isn't a good way to incorporate such weird format with an ability to have winter dates crossing next year treshold. Therefore I'm removing ability to add them like that, since the same thing can be accomplished by writing for example 20.12-26.12