Fix rare crash when changing NPC graphics
The crash might have happened when the original graphic has more frames than the new graphic, and the current active frame when changing it is higher than the number of new frames. This would make it so the game would try to access frames out of bounds, and thus crashing the draw pipeline.