Change the build system to be more straightforward
This commit adds an instruction and workflow for a quick setup for developing the addon.
The userscript is now generated from the .env file,
so it is ready to be used for local development.

Webpack now also rebuilds the app when the files the app requires change (version, npc or lights).