Change hiding chat commercials to be a setting turned off by default
It can't be turned on by default due to some "commercials" being
actually useful information like an incoming maintenance break.
Hiding such information might lead to disruptions in the regular game.
(For example,
trying to fight a hard mob right before a maintenance break
could lead to a player being kicked out in the middle of the fight.)