Add prettier to the project
The change in code style may be a bit of a shock. However, during the years, prettier became the de-facto standard way of formatting the JS/TS projects. We could try to change the options to better match the previous codebase, but I think it would be easier for newcomers if we embrace the defaults, like the prettier docs suggest.