Optimize updating NPCs with custom GIF images by using canvas.toBlob() instead of canvas.toDataURL()
While the performance hit of having the addon installed should be close to zero,
on a map with a lot of NPCs with changed graphics, the performance hit was still insignificant, but measurably worse.
This particular optimization makes loading the addon 41% faster in extreme cases (from 67 ms to 44 ms) which makes about a 2-3% difference in total time spent when loading the game.