Replace FakeNpc with Drawable
This implementation should be more resilient to game updates,
since we don't create broken, fake versions of NPCs,
but instead use an appropriate API for custom renderings.

This wasn't implemented before,
because either the API did not exist
or its existence was missed
when scouting for a potential hook to render custom stuff on the new interface.